+++ app/ui/forecast/topnav.tsx
@@ -1,0 +1,17 @@
+import Link from 'next/link';
+import NavLinks from '@/app/ui/forecast/nav-links';
+import AcmeLogo from '@/app/ui/acme-logo';
+import { PowerIcon } from '@heroicons/react/24/outline';
+
+export default function TopNav() {
+  return (
+ 
+    <div className="flex grow flex-row justify-between space-x-2 md:flex-col md:space-x-0 md:space-y-2">
+          <NavLinks />
+        <div className="hidden h-auto w-full grow rounded-md bg-gray-50 md:block"></div>
+    </div>
+    
+
+
+  );
+}